Extract `IntegrationTest.kt`
