+++ buildSrc/src/main/kotlin/io/spine/internal/gradle/javascript/task/Check.kt
@@ -69,7 +69,6 @@ fun JsTaskRegistering.check(configuration: JsTaskRegistering.() -> Unit = {}) {
 }
 
 
-
 internal val TaskContainer.checkJs: Task
     get() = getByName("checkJs")
